refactor : 역할에 따른 클래스 분리
